+++ tesla_http_proxy/CHANGELOG.md
@@ -1,5 +1,17 @@
 <!-- https://developers.home-assistant.io/docs/add-ons/presentation#keeping-a-changelog -->
 
+## 2.2.5
+
+### Changed
+
+- Retry public key check if HTTP status code != 200
+
+## 2.2.4
+
+### Changed
+
+- Retry public key check if no IP address is found for the FQDN
+
 ## 2.2.3
 
 ### Changed